Wire JobKorea talent helper into CI
